requirements.txt
pillow
django

# copied from https://github.com/point97/madrona/blob/master/requirements.txt
boto
django-compressor
django-celery
django-flatblocks
django-kombu
django-picklefield
django-registration
#-e git://github.com/springmeyer/djmapnik.git#egg=djmapnik
#djmapnik
numpy
psycopg2
python-openid
#South>=0.7
#wsgiref

django-tastypie
django-extensions
-e git+https://github.com/Ecotrust/madrona@py3#egg=madrona
simplejson
django-redis
querystring-parser
httplib2
requests
oauthlib
gunicorn
django-proxy
MapProxy
eventlet
ipython
python-social-auth

geopy
django-tinymce
django-wysiwyg
